backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic[email]>=2.4.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
# psycopg2-binary==2.9.9  # Commented out as it requires PostgreSQL
alembic>=1.12.1
python-dotenv>=1.0.0
# Add MySQL driver
pymysql>=1.1.0